requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
boto3==1.17.68
    # via eventhooks
botocore==1.20.68
    # via
    #   boto3
    #   s3transfer
certifi==2020.12.5
    # via requests
chardet==4.0.0
    # via requests
eventhooks[aws]==0.2.5
    # via -r requirements.in
idna==2.10
    # via requests
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
python-dateutil==2.8.1
    # via botocore
pyyaml==5.4.1
    # via -r requirements.in
requests==2.25.1
    # via eventhooks
s3transfer==0.4.2
    # via boto3
six==1.16.0
    # via python-dateutil
urllib3==1.26.4
    # via
    #   botocore
    #   requests
